refactor: address daily quality findings for 2026-08-16
Split Git History panel chrome/SVG/row modules, extract automation setup
leave-guard hook, reuse Skills detail URL helper, and archive the quality
report (score 79).
